requirements.txt
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
xgboost==2.0.2
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
yfinance==0.2.28
requests==2.31.0
python-dotenv==1.0.0
ta==0.10.2
vectorbt==0.25.2
joblib==1.3.2
streamlit==1.29.0
schedule==1.2.0
pyyaml==6.0.1

